requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
flask_cors >= 3.0.8
mysql-connector >= 2.2.9
scipy
